% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VINTAGE.R
\name{run_vintage}
\alias{run_vintage}
\title{Run VINTAGE}
\usage{
run_vintage(
  eqtl,
  gwas,
  refpath,
  gene_info,
  outfile,
  ref_type = c("plink"),
  ambiguity = TRUE,
  maf = 0.05,
  window = 1e+05,
  B = 1e+06,
  dofilter = TRUE,
  pg_cutoff = 0.05/20000,
  ncores = 1,
  maxIterEM = 1e+05,
  tolEM = 0.001,
  save_profile = 1,
  seed = 0
)
}
\arguments{
\item{eqtl}{eQTL summary statistics}

\item{gwas}{GWAS summary statistics}

\item{refpath}{Reference panel for calculating SNP-SNP correlation}

\item{gene_info}{Gene annotations}

\item{outfile}{Output file name}

\item{ref_type}{Format of reference panel}

\item{ambiguity}{Whether to filter out strand ambiguous variants}

\item{maf}{Filter out variants with minor allele frequency below this value}

\item{window}{Integer. Buffer region size flanking the gene TSS and TES}

\item{B}{Number of simulations to evaluate the simulation-based p-value}

\item{dofilter}{Whether to filter out LD mismatched variants}

\item{pg_cutoff}{Perform genetic correlation test if the p-value of the
genetic variance test passed pg_cutoff}

\item{ncores}{Number of cores to use}

\item{maxIterEM}{Maximum number of iterations for the EM algorithm}

\item{tolEM}{Tolerance for likelihood difference to claim convergence}

\item{save_profile}{Evaluate likelihood}

\item{seed}{Random seed}
}
\description{
VINTAGE analyzes genes on one chromosome at a time
}
