\name{VINTAGE-package}
\alias{VINTAGE-package}
\alias{VINTAGE}
\docType{package}
\title{\packageTitle{VINTAGE}}
\description{\packageDescription{VINTAGE}}
\section{Package Content}{\packageIndices{VINTAGE}}
\author{\packageAuthor{VINTAGE}}
\section{Maintainer}{\packageMaintainer{VINTAGE}}
\keyword{package}
