\name{VINTAGE-package}
\alias{VINTAGE-package}
\alias{VINTAGE}
\docType{package}
\title{
  A unified framework integrating gene expression mapping studies with 
  genome-wide association studies for detecting and deciphering gene-trait 
  associations
}
\description{
  This package implements the VINTAGE method for an integrative analysis of
  gene expression mapping studies and GWAS to identify and decipher gene-trait
  associations.
}
\section{Package Content}{\packageIndices{VINTAGE}}
\author{
  Zheng Li, Boran Gao, Xiang Zhou
}
\section{Maintainer}{
  Zheng Li <zlisph@umich.edu>
}
\keyword{package}
